backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1